+++ HISTORY.md
@@ -9,3 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.1] - 2022-04-19
 * Sorry, pushed copy/pasted changelog on first release!
+
+## [0.2.0] - 2022-08-11
+* Merged `HIDE_APPS` and `HIDE_PERMS` to make more clear that we're hiding perms only.
+* Added ability to extend your own custom `UserAdmin` and `GroupAdmin` if need be.